pyproject.toml
[project]
name = "stock-master"
version = "0.1.0"
description = "本地优先的个人股票投资决策系统：数据驱动 + 多模型调研 + 人类裁决"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare",
    "pandas",
    "typer",
    "rich",
    "pyyaml",
    "pydantic>=2.0",
    "jinja2",
]

[project.scripts]
sm = "stock_master.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_master"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
